+++ src/cuda/3d/interp3d_wrapper.cu
@@ -29,7 +29,7 @@ int cuinterp3d(cufinufft_plan_t<T> *d_plan, int blksize)
     int nf1 = d_plan->nf1;
     int nf2 = d_plan->nf2;
     int nf3 = d_plan->nf3;
-    int M = d_plan->M;
+    int64_t M = d_plan->M;
 
     int ier;
     switch (d_plan->opts.gpu_method) {
@@ -48,7 +48,7 @@ int cuinterp3d(cufinufft_plan_t<T> *d_plan, int blksize)
 }
 
 template <typename T>
-int cuinterp3d_nuptsdriven(int nf1, int nf2, int nf3, int M, cufinufft_plan_t<T> *d_plan, int blksize) {
+int cuinterp3d_nuptsdriven(int nf1, int nf2, int nf3, int64_t M, cufinufft_plan_t<T> *d_plan, int blksize) {
     auto &stream = d_plan->stream;
 
     dim3 threadsPerBlock;
@@ -93,7 +93,7 @@ int cuinterp3d_nuptsdriven(int nf1, int nf2, int nf3, int M, cufinufft_plan_t<T>
 }
 
 template <typename T>
-int cuinterp3d_subprob(int nf1, int nf2, int nf3, int M, cufinufft_plan_t<T> *d_plan, int blksize) {
+int cuinterp3d_subprob(int nf1, int nf2, int nf3, int64_t M, cufinufft_plan_t<T> *d_plan, int blksize) {
     auto &stream = d_plan->stream;
 
     int ns = d_plan->spopts.nspread; // psi's support in terms of number of cells
@@ -155,13 +155,13 @@ int cuinterp3d_subprob(int nf1, int nf2, int nf3, int M, cufinufft_plan_t<T> *d_
 template int cuinterp3d<float>(cufinufft_plan_t<float> *d_plan, int blksize);
 template int cuinterp3d<double>(cufinufft_plan_t<double> *d_plan, int blksize);
 
-template int cuinterp3d_nuptsdriven<float>(int nf1, int nf2, int nf3, int M, cufinufft_plan_t<float> *d_plan,
+template int cuinterp3d_nuptsdriven<float>(int nf1, int nf2, int nf3, int64_t M, cufinufft_plan_t<float> *d_plan,
                                            int blksize);
-template int cuinterp3d_nuptsdriven<double>(int nf1, int nf2, int nf3, int M, cufinufft_plan_t<double> *d_plan,
+template int cuinterp3d_nuptsdriven<double>(int nf1, int nf2, int nf3, int64_t M, cufinufft_plan_t<double> *d_plan,
                                             int blksize);
 
-template int cuinterp3d_subprob<float>(int nf1, int nf2, int nf3, int M, cufinufft_plan_t<float> *d_plan, int blksize);
-template int cuinterp3d_subprob<double>(int nf1, int nf2, int nf3, int M, cufinufft_plan_t<double> *d_plan,
+template int cuinterp3d_subprob<float>(int nf1, int nf2, int nf3, int64_t M, cufinufft_plan_t<float> *d_plan, int blksize);
+template int cuinterp3d_subprob<double>(int nf1, int nf2, int nf3, int64_t M, cufinufft_plan_t<double> *d_plan,
                                         int blksize);
 
 } // namespace spreadinterp
